+++ test/process/process.ts
@@ -163,7 +163,7 @@ export const IsUser = (duplo: DuploInstance<DuploConfig>) => {
 		isOdd,
 		{
 			input: p => 2,
-			result: "odd",
+			result: ["odd"],
 			catch: (res, i) => res.info(i).code(400).send("wrong"),
 			indexing: "result",
 			options: p => ({
